Add tx flows to new modal context
