+++ src/components/TxFlow/TokenTransfer/TokenTransferFlow.tsx
@@ -1,0 +1,37 @@
+import { TokenTransferStepper } from '@/components/TxFlow/TokenTransfer/index'
+import { SendTxType } from '@/components/tx/modals/TokenTransferModal/SendAssetsForm'
+import CreateTokenTransfer from '@/components/TxFlow/TokenTransfer/CreateTokenTransfer'
+import ReviewTokenTransfer from '@/components/TxFlow/TokenTransfer/ReviewTokenTransfer'
+import TxLayout from '@/components/TxFlow/common/TxLayout'
+import { ProgressBar } from '@/components/common/ProgressBar'
+
+const initialData = {
+  recipient: '',
+  tokenAddress: '',
+  amount: '',
+  type: SendTxType.multiSig,
+}
+
+const TokenTransferFlow = ({ txNonce }: { txNonce?: number }) => {
+  const steps = [
+    <CreateTokenTransfer key={0} disableSpendingLimit={txNonce !== undefined} />,
+    <ReviewTokenTransfer key={1} txNonce={txNonce} />,
+  ]
+
+  return (
+    <TokenTransferStepper.Provider steps={steps} defaultValues={[initialData, {}]}>
+      {(Step, values) => {
+        return (
+          <TxLayout title="Send Tokens">
+            <>
+              <ProgressBar value={values.progress} />
+              {Step}
+            </>
+          </TxLayout>
+        )
+      }}
+    </TokenTransferStepper.Provider>
+  )
+}
+
+export default TokenTransferFlow
