+++ src/components/common/ModalDialog/index.tsx
@@ -1,4 +1,4 @@
-import { type ReactElement, type ReactNode } from 'react'
+import { type ReactElement, type ReactNode, useContext } from 'react'
 import { IconButton, type ModalProps } from '@mui/material'
 import { Dialog, DialogTitle, type DialogProps, useMediaQuery } from '@mui/material'
 import { useTheme } from '@mui/material/styles'
@@ -6,6 +6,7 @@ import ChainIndicator from '@/components/common/ChainIndicator'
 import CloseIcon from '@mui/icons-material/Close'
 
 import css from './styles.module.css'
+import { ModalContext } from '@/components/TxFlow/ModalProvider'
 
 interface ModalDialogProps extends DialogProps {
   dialogTitle?: React.ReactNode
@@ -51,22 +52,50 @@ const ModalDialog = ({
   ...restProps
 }: ModalDialogProps): ReactElement => {
   const theme = useTheme()
-  const isSmallScreen = useMediaQuery(theme.breakpoints.down('sm'))
-  const isFullScreen = fullScreen || isSmallScreen
+  const isSmallScreen = useMediaQuery(theme.breakpoints.down('md'))
+  const { setVisibleModal } = useContext(ModalContext)
 
   return (
     <Dialog
       {...restProps}
-      fullScreen={isFullScreen}
+      fullScreen={true}
       scroll={fullScreen ? 'paper' : 'body'}
       className={css.dialog}
       onClick={(e) => e.stopPropagation()}
+      sx={{
+        top: 52,
+        left: isSmallScreen ? 0 : 230,
+        zIndex: 1200,
+      }}
+      slots={{ backdrop: 'div' }}
+      PaperProps={{
+        sx: {
+          pb: 8,
+          backgroundColor: (theme) => theme.palette.border.background,
+        },
+      }}
     >
-      {dialogTitle && (
-        <ModalDialogTitle onClose={restProps.onClose} hideChainIndicator={hideChainIndicator}>
-          {dialogTitle}
-        </ModalDialogTitle>
-      )}
+      <DialogTitle
+        sx={{
+          mb: 8,
+          display: 'flex',
+          alignItems: 'center',
+          backgroundColor: (theme) => theme.palette.background.paper,
+        }}
+      >
+        <span style={{ flex: 1 }} />
+        <IconButton
+          aria-label="close"
+          onClick={() => setVisibleModal(undefined)}
+          size="small"
+          sx={{
+            ml: 2,
+            color: 'border.main',
+          }}
+        >
+          <CloseIcon />
+        </IconButton>
+      </DialogTitle>
 
       {children}
     </Dialog>
