+++ src/components/sidebar/NewTxButton/index.tsx
@@ -1,45 +1,35 @@
-import { Suspense, useState, type ReactElement } from 'react'
-import dynamic from 'next/dynamic'
+import { type ReactElement, useContext } from 'react'
 import Button from '@mui/material/Button'
 import css from './styles.module.css'
-import { trackEvent, OVERVIEW_EVENTS } from '@/services/analytics'
+import { OVERVIEW_EVENTS, trackEvent } from '@/services/analytics'
 import CheckWallet from '@/components/common/CheckWallet'
-
-const NewTxModal = dynamic(() => import('@/components/tx/modals/NewTxModal'))
+import { ModalContext, ModalType } from '@/components/TxFlow/ModalProvider'
 
 const NewTxButton = (): ReactElement => {
-  const [txOpen, setTxOpen] = useState<boolean>(false)
+  const { setVisibleModal } = useContext(ModalContext)
 
   const onClick = () => {
-    setTxOpen(true)
+    setVisibleModal({ type: ModalType.NewTx, props: {} })
 
     trackEvent(OVERVIEW_EVENTS.NEW_TRANSACTION)
   }
 
   return (
-    <>
-      <CheckWallet allowSpendingLimit>
-        {(isOk) => (
-          <Button
-            onClick={onClick}
-            variant="contained"
-            size="small"
-            disabled={!isOk}
-            fullWidth
-            className={css.button}
-            disableElevation
-          >
-            New transaction
-          </Button>
-        )}
-      </CheckWallet>
-
-      {txOpen && (
-        <Suspense>
-          <NewTxModal onClose={() => setTxOpen(false)} />
-        </Suspense>
+    <CheckWallet allowSpendingLimit>
+      {(isOk) => (
+        <Button
+          onClick={onClick}
+          variant="contained"
+          size="small"
+          disabled={!isOk}
+          fullWidth
+          className={css.button}
+          disableElevation
+        >
+          New transaction
+        </Button>
       )}
-    </>
+    </CheckWallet>
   )
 }
