+++ src/components/transactions/RejectTxButton/index.tsx
@@ -1,10 +1,9 @@
 import type { TransactionSummary } from '@safe-global/safe-gateway-typescript-sdk'
 import { Button, SvgIcon, Tooltip } from '@mui/material'
 
-import type { ReactElement, SyntheticEvent } from 'react'
-import { Suspense, useContext, useState } from 'react'
+import type { ReactElement } from 'react'
+import { useContext } from 'react'
 import { isMultisigExecutionInfo } from '@/utils/transaction-guards'
-import dynamic from 'next/dynamic'
 import useIsPending from '@/hooks/useIsPending'
 import IconButton from '@mui/material/IconButton'
 import ErrorIcon from '@/public/images/notifications/error.svg'
@@ -13,9 +12,7 @@ import { TX_LIST_EVENTS } from '@/services/analytics/events/txList'
 import CheckWallet from '@/components/common/CheckWallet'
 import { useSafeSDK } from '@/hooks/coreSDK/safeCoreSDK'
 import { getTxButtonTooltip } from '@/components/transactions/utils'
-import { ModalContext, ModalType } from '@/services/ModalProvider'
-
-const NewTxModal = dynamic(() => import('@/components/tx/modals/NewTxModal'))
+import { ModalContext, ModalType } from '@/components/TxFlow/ModalProvider'
 
 const RejectTxButton = ({
   txSummary,
@@ -25,7 +22,6 @@ const RejectTxButton = ({
   compact?: boolean
 }): ReactElement | null => {
   const { setVisibleModal } = useContext(ModalContext)
-  const [open, setOpen] = useState<boolean>(false)
   const txNonce = isMultisigExecutionInfo(txSummary.executionInfo) ? txSummary.executionInfo.nonce : undefined
   const isPending = useIsPending(txSummary.id)
   const safeSDK = useSafeSDK()
@@ -33,49 +29,30 @@ const RejectTxButton = ({
 
   const tooltipTitle = getTxButtonTooltip('Replace', { hasSafeSDK: !!safeSDK })
 
-  const onClick = (e: SyntheticEvent) => {
-    e.stopPropagation()
-    setOpen(true)
+  const openReplacementModal = () => {
+    setVisibleModal({ type: ModalType.ReplaceTx, props: { txNonce } })
   }
 
   return (
-    <>
-      <CheckWallet>
-        {(isOk) => (
-          <Track {...TX_LIST_EVENTS.REJECT}>
-            {compact ? (
-              <Tooltip title={tooltipTitle} arrow placement="top">
-                <span>
-                  <IconButton
-                    onClick={() => setVisibleModal({ type: ModalType.ReplaceTx, props: { txNonce, open: true } })}
-                    color="error"
-                    size="small"
-                    disabled={!isOk || isDisabled}
-                  >
-                    <SvgIcon component={ErrorIcon} inheritViewBox fontSize="small" />
-                  </IconButton>
-                </span>
-              </Tooltip>
-            ) : (
-              <Button
-                onClick={() => setVisibleModal({ type: ModalType.ReplaceTx, props: { txNonce, open: true } })}
-                variant="danger"
-                disabled={!isOk || isDisabled}
-                size="stretched"
-              >
-                Replace
-              </Button>
-            )}
-          </Track>
-        )}
-      </CheckWallet>
-
-      {open && (
-        <Suspense>
-          <NewTxModal onClose={() => setOpen(false)} txNonce={txNonce} />
-        </Suspense>
+    <CheckWallet>
+      {(isOk) => (
+        <Track {...TX_LIST_EVENTS.REJECT}>
+          {compact ? (
+            <Tooltip title={tooltipTitle} arrow placement="top">
+              <span>
+                <IconButton onClick={openReplacementModal} color="error" size="small" disabled={!isOk || isDisabled}>
+                  <SvgIcon component={ErrorIcon} inheritViewBox fontSize="small" />
+                </IconButton>
+              </span>
+            </Tooltip>
+          ) : (
+            <Button onClick={openReplacementModal} variant="danger" disabled={!isOk || isDisabled} size="stretched">
+              Replace
+            </Button>
+          )}
+        </Track>
       )}
-    </>
+    </CheckWallet>
   )
 }
